// DoubleVPartnersFrontEnd/src/app/models/persona-model.ts
export class Persona {
  Identificador?: number;
  Nombres?: string;
  Apellidos?: string;
  NumeroIdentificacion?: string;
  Email?: string;
  TipoIdentificacion?: string;
  FechaCreacion?: string;
  NumeroIdentificacionConcatenado?: string;
  NombresApellidosConcatenados?: string;
  UsuarioId?: number;
}